[pref] `drive list` also prints local Drives; `drive set` for setup or modify drive config.

+++ onedrived/od_context.py
@@ -5,7 +5,7 @@ from pwd import getpwnam
 
 import xdg
 
-from . import mkdir
+from . import mkdir, get_resource
 from .models import account_profile
 from .models import drive_config
 
@@ -33,6 +33,7 @@ class UserContext:
     }
 
     DEFAULT_CONFIG_FILENAME = 'onedrived_config_v2.json'
+    DEFAULT_IGNORE_FILENAME = 'ignore_v2.txt'
 
     SUPPORTED_PROXY_PROTOCOLS = ('http', 'https')
 
@@ -53,20 +54,23 @@ class UserContext:
         self.config = self.DEFAULT_CONFIG
 
     def _create_config_dir_if_missing(self):
+        if os.path.exists(self.config_dir) and not os.path.isdir(self.config_dir):
+            self.logger.info('Config dir "' + self.config_dir + '" is not a directory. Delete it.')
+            os.remove(self.config_dir)
         if not os.path.exists(self.config_dir):
             self.logger.info('Config dir "' + self.config_dir + '" does not exist. Create it.')
             mkdir(self.config_dir, self.user_uid, mode=0o700, exist_ok=True)
-        elif not os.path.isdir(self.config_dir):
-            self.logger.info('Config dir "' + self.config_dir + '" is not a directory. Delete and re-create it.')
-            os.remove(self.config_dir)
-            mkdir(self.config_dir, self.user_uid, mode=0o700, exist_ok=True)
+            with open(self.config_dir + '/' + self.DEFAULT_IGNORE_FILENAME, 'w') as f:
+                f.write(get_resource('data/ignore_v2.txt'))
 
     def add_account(self, account_profile):
         """
         Add a new account to config file.
         :param models.account_profile.OneDriveAccountProfile account_profile:
+        :return models.account_profile.OneDriveAccountProfile: The account profile argument.
         """
         self.config['accounts'][account_profile.account_id] = account_profile.data
+        return account_profile
 
     def get_account(self, account_id):
         """
@@ -91,15 +95,17 @@ class UserContext:
         """
         Add a new drive to local config.
         :param models.drive_config.LocalDriveConfig drive_config:
+        :return models.drive_config.LocalDriveConfig drive_config:
         """
         self.config['drives'][drive_config.drive_id] = drive_config.data
+        return drive_config
 
     def get_drive(self, drive_id):
         """
         :param str drive_id:
         :return models.drive_config.LocalDriveConfig:
         """
-        return drive_config.LocalDriveConfig(drive_id, **self.config['drives'][drive_id])
+        return drive_config.LocalDriveConfig(**self.config['drives'][drive_id])
 
     def delete_drive(self, drive_id):
         del self.config['drives'][drive_id]
